ceph-volume: fix list command failing on volumes sharing the same name
Given a system with the following disk setup:

    nvme1n1               259:0    0   1.8T  0 disk
    └─group1-volume       253:1    0   1.8T  0 lvm
    nvme0n1               259:0    0   1.8T  0 disk
    └─group2-volume       253:2    0   1.8T  0 lvm

That is, two disks; the first one with a vg named "group1", and in that vg
an lv named "volume"; and the second one with a vg named "group2" and in
that vg an lv named "volume".

`ceph-volume lvm list group1/volume` would fail with the following error:

    $ ceph-volume lvm list group1/volume
    -->  RuntimeError: Filters {'lv_name': 'volume'} matched more than 1 LV present on this host.

... even though there should be no ambiguity. The command contains both
the vg name and the lv name, yet the filter that was used only
considered lv_name for some reason.

This caused a lot of confusion as it caused rook-ceph to fail to
provision OSDs on a machine with a similar setup, because rook was
calling `ceph-volume lvm list --format=json vg/lv`, and getting a
nonzero exit status.

Signed-off-by: Franklin "Snaipe" Mathieu <me@snai.pe>