rbd-mirror: fix acquire leader lock to avoid replayer stuck in stopped state
execute_timer_task()
  -> acquire_leader_lock()
    -> handle_acquire_leader_lock: r=-110

In case of a released leader lock, the image replayer will be in stopped state,
and incase if an attempt to acquire the leader lock failed with ETIMEDOUT,
it leads to image replayer stuck in stopped state forever. This is
because any reattempt to acquire the lock fails with "error acquiring lock:
 (17) File exists"

Also currently `rbd_mirror_leader_max_acquire_attempts_before_break` is
set to 3 by default once m_acquire_attempts reaches the limit,
break_leader_lock will be invoked. But in this case the break_leader_lock fails
with "error breaking leader lock: (22) Invalid argument", this might be because
the lock owner is still alive.

This way, the image replayer stays in stopped state forever.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>