cmake: remove versioning from cls_* dynamic libraries to avoid runtime error
Darwin dynamic library versioning is injecting the numbers before the suffix, like this:

  cls_journal.1.dylib

Such naming is not handled by the ClassHandler and it attempts to load a library with the name `journal.1`
The code proceeds initializing the library which calls to register but with the name `journal`
which doesn't match the cached class list entry. Registration fails and the library proceeds
using the NULL handle, crashing the osd process.

Signed-off-by: Leonid Usov <leonid.usov@ibm.com>