common/TrackedOp: fix race updating description with proper lock
It is not safe to "cache" a member which may be changed by a racing thread.

This reworks the locking so we can do a light-weight check if the description
is already generated wihtout acquiring the heavier TrackedOp::lock. If it's not
available yet or it needs regenerated, then get the proper locks to generate
it.

Fixes: b4cf6784b5e739c4941d30de16509df31e411a86
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>