common/TrackedOp: support overriding the _dump method
When dumping an op, it may be desirable to alter how it is dumped depending on
which locks are held. As it happens, I plan to dump extra information if the
mds_lock is held!

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>
(cherry picked from commit 224ba5ad4cf7cce6f48d9cbe1e0157b804954c6a)