doc: clarify upmap balancer documentation
In https://github.com/ceph/ceph/pull/51499, some language was added
that refers to the upmap balancer as the "capacity" balancer,
but we don't want to insinuate that there is anything "new" that was
added or changed. It should be clear that this section refers to the
existing upmap balancer, which has existed for several releases.

Signed-off-by: Laura Flores <lflores@redhat.com>