osd: check adjacent clones are readable when rollback is called
When an object executes a rollback op from its clone, it can bring this issue
if any of the adjacent clones of rollback destination is unreadable.

To fix this issue, let OSD check if the rollback destination's adjacent clones are readable.
If they are unreadable, wait until they are available.

fixes: https://tracker.ceph.com/issues/62167

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>
Signed-off-by: Sungmin Lee <sung_min.lee@samsung.com>