qa/cephadm: mgr-nfs-upgrade, match any migration > 2
I believe this check was originally added because
the 2->3 migration migrated some nfs related bits. Since
then we've had to update the migration this checks
for every time we bump the max migration. This change
is intended to instead just have it check for a
miration > 2 so we don't have to keep updating it.

Signed-off-by: Adam King <adking@redhat.com>