mgr/dashboard: automate centralized logging configuration
Configure screen when either loki, promtail or grafana is not
configured.
Configure button which
- Creates grafana/loki/promtail daemons
- Set the config options needed

The nav tabs in the logs is now routerized.

The filter toolbar is now hidden for centralized logging because it has
no use there. This will also increase the height of the centralized log
card.

The Daemon Logs is now renamed to be `Centralized`.

Added an option to configure the error component to have breadcrumbs and
nav-tab so there won't be any visual differences between the other log pages
when centralized logging is not configured.

made the error page icons a bit more smaller.
Configure button will now have a spinner to indicate the configuration
is under progress.

Fixes: https://tracker.ceph.com/issues/63240
Signed-off-by: Nizamudeen A <nia@redhat.com>