mds/linkmerge: Remove referent inode on linkmerge
When the primary file is remove before the secondary,
stray reintegration makes any secondary link available
in the cache to be primary. This is referred as linkmerge.

When the linkmerge happens, the referent inode of the
secondary needs to be removed otherwise it causes the
referent inode object leak in the data pool. This PR
handles the linkmerge scenario and removed the referent
inode of the secondary during the linkmerge.

Signed-off-by: Kotresh HR <khiremat@redhat.com>