os/bluestore: replicate on mkfs/mount, read replicated always
Instead of replicating when writing running `_write_bdev_label` we
replicate on mkfs and in mount.

`_replicate_bdev_label` will try to replicate the label by checking
first if the allocator didn't allocate anything in the offset we want to
replicate at.

`read_bdev_label` will read all offsets where we might find labels as it
expects at least one to be ok. If it wasn't replicated and we try to
read other offset a part from "0", we assume it will fail on crc check.

Signed-off-by: Pere Diaz Bou <pere-altea@hotmail.com>