mgr/nfs: do not suppress errors while updating export
for any failure while updating export, it reports error like this:

$ ceph nfs export apply my-nfs -i export.conf
[
  {
    "msg": "Failed to apply export: export FSAL user_id must be 'nfs.my-nfs.1'",
    "state": "error"
  }
]
Error EIO:

This makes it always return code 5, no matter what the actual errno was/should
be. This must be changed to report the correct errno, and a single line
containing error message instead of a dict, i.e. with this patch it would be
like this:

$ ceph nfs export apply my-nfs -i export.conf
Error EINVAL: export FSAL user_id must be 'nfs.my-nfs.1'

This looks much better, neat and correct IMO.

Signed-off-by: Dhairya Parmar <dparmar@redhat.com>