rgw/d4n: BlockDirectory dispatches to connection executor
adds an async_exec() wrapper that initiates the call to
connection::async_exec() on that connection's executor. this was
necessary to allow calls to redis_exec() from arbitrary threads

conn_cancel() (renamed to shutdown) also has to dispatch its call to
cancel() to the connection's executor

to prevent races between shutdown()'s asynchronous use of dispatch and
BlockDirectory's destruction, the boost::redis::connection is managed by
a std::shared_ptr. this allows the connection's lifetime to be extended
by async_exec() and shutdown() even if the BlockDirectory is destroyed

Signed-off-by: Casey Bodley <cbodley@redhat.com>