mgr/dashboard: remove green tick on old password field
a green tick is showing to the field where we enter the old password in
login password change form. It starts showing green tick as soon as we
start typing on it. Removing that because its misleads the user.

Fixes: https://tracker.ceph.com/issues/62644
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit 798a95dece66095db8ef0de9ed81c99fe23c4a11)