cephadm: remove gateways property from Ceph type
The gateways list property of the Ceph type was used in exactly one
place. In order to provide a clearer structure to the classes in
cephadm, remove this property and simply list the daemon types
specifically where they are needed.

In the future, I hope to see these handled in a class/object based
manner but this is still better for now.

Signed-off-by: John Mulligan <jmulligan@redhat.com>