mgr/balancer: Improved 'ceph balancer status' output
We have recently introduced a detailed command to check the status of our balancer.
This command displays the location of pg_up_map items where they are being transferred
when the balancer runs. The primary objective of this command is to obtain a better
understanding of the balancer's behavior in terms of transferring pgs. This will help
us identify when the balancing is taking place in the effort to improve the balancing
process to evenly distribute pgs in the balancing process.

Signed-off-by: Christopher Poon <poon3@rpi.edu>
Signed-off-by: Zuriel Aviles <avilez@rpi.edu>
Signed-off-by: Nico Wang <wangn4@rpi.edu>